+++ src/main/java/mk/ukim/finki/application/customer/CustomerDao.java
@@ -6,15 +6,15 @@ import java.util.Optional;
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
 
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
 
     Customer saveCustomer(Customer customer);
 
     boolean existsCustomerWithEmail(String email);
 
-    void deleteCustomerById(Integer id);
+    void deleteCustomerById(Long id);
 
-    boolean existsCustomerWithId(Integer id);
+    boolean existsCustomerWithId(Long id);
 
     Customer updateCustomer(Customer customer);
 }
